Make PID test Linux-only
